// portal/proto/isuxportal/services/admin/teams_pb.d.ts
export declare type ListTeamsResponse = Message<"isuxportal.proto.services.admin.ListTeamsResponse"> & {
  /**
   * @generated from field: repeated isuxportal.proto.services.admin.ListTeamsResponse.TeamListItem teams = 1;
   */
  teams: ListTeamsResponse_TeamListItem[];
};

/**
 * Describes the message isuxportal.proto.services.admin.ListTeamsResponse.
 * Use `create(ListTeamsResponseSchema)` to create a new message.
 */
export declare const ListTeamsResponseSchema: GenMessage<ListTeamsResponse>;

/**
 * @generated from message isuxportal.proto.services.admin.ListTeamsResponse.TeamListItem
 */
export declare type ListTeamsResponse_TeamListItem = Message<"isuxportal.proto.services.admin.ListTeamsResponse.TeamListItem"> & {
  /**
   * @generated from field: int64 team_id = 1;
   */
  teamId: bigint;

  /**
   * @generated from field: string name = 2;
   */
  name: string;

  /**
   * @generated from field: repeated string member_names = 3;
   */
  memberNames: string[];

  /**
   * @generated from field: bool final_participation = 4;
   */
  finalParticipation: boolean;

  /**
   * @generated from field: bool is_student = 5;
   */
  isStudent: boolean;

  /**
   * @generated from field: bool withdrawn = 6;
   */
  withdrawn: boolean;

  /**
   * @generated from field: bool disqualified = 7;
   */
  disqualified: boolean;

  /**
   * @generated from field: bool hidden = 8;
   */
  hidden: boolean;
};